ml-intelligence/requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.34
joblib==1.3.2
psycopg2-binary==2.9.9
fastapi==0.115.0
uvicorn==0.30.0